Atualização modelagem

+++ database/migrations/2022_11_29_151959_create_candidatos_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('estado', 50);
             $table->string('filiação', 50);
             $table->string('número', 50)->unique();
-            $table->bigInteger('votos');
+            $table->integer('votos');
             $table->boolean('situação'); 
             $table->timestamps();
             $table->softDeletes();
